requirements.txt
flask
openai-whisper
